feat server validation, refactor serverActions

+++ src/app/actions/upvoteActions.ts
@@ -4,7 +4,7 @@ import { createUpvote, deleteUpvote, getUpvoters } from '@/db/queries/upvotes';
 import { revalidatePath, revalidateTag } from 'next/cache';
 import { Upvote } from '@/types';
 import { isValidUUID } from '@/lib/utils';
-import { auth } from '../auth';
+import { validateAuthorization } from '../auth';
 
 export async function getUpvotersAction(feedbackId: string): Promise<Upvote[]> {
   if (!isValidUUID(feedbackId)) {
@@ -26,10 +26,7 @@ export async function createUpvoteAction(feedbackId: string, userId: string): Pr
   if (!isValidUUID(feedbackId) || !isValidUUID(userId)) {
     throw new Error('Invalid feedbackId or userId');
   }
-  const session = await auth();
-  if (!session) {
-    throw new Error('Not authorized in createCommentAction');
-  }
+  await validateAuthorization('createUpvoteAction', userId);
   try {
     await createUpvote(feedbackId, userId);
     revalidateTag('feedbacks');
@@ -44,7 +41,7 @@ export async function deleteUpvoteAction(feedbackId: string, userId: string): Pr
   if (!isValidUUID(feedbackId) || !isValidUUID(userId)) {
     throw new Error('Invalid feedbackId or userId');
   }
-
+  await validateAuthorization('deleteUpvoteAction', userId);
   try {
     await deleteUpvote(feedbackId, userId);
     revalidateTag('feedbacks');
